+++ packages/terra-slide-panel/src/terra-dev-site/doc/example/DefaultSlidePanel.jsx
@@ -21,7 +21,7 @@ const mainContentForSlidePanel = togglePanelHandler => (
       <p>
         {'Focus is moved to the toggle button in the panel container when the panel is opened via the componentDidUpdate lifecycle hook in '}
         <a href="https://github.com/cerner/terra-framework/blob/main/packages/terra-slide-panel/src/terra-dev-site/doc/example/DefaultSlidePanel.jsx">the example code</a>
-        {'.'}
+        .
       </p>
       <ul>
         <li>Item 1</li>
@@ -49,7 +49,7 @@ const panelContentForSlidePanel = (togglePanelHandler, toggleFullscreenHandler) 
       <p>
         {'Focus is moved to the toggle button in the main container when the panel is closed via the componentDidUpdate lifecycle hook in '}
         <a href="https://github.com/cerner/terra-framework/blob/main/packages/terra-slide-panel/src/terra-dev-site/doc/example/DefaultSlidePanel.jsx">the example code</a>
-        {'.'}
+        .
       </p>
       <ul>
         <li>Item 1</li>