+++ packages/terra-date-input/src/terra-dev-site/test/date-input/FocusBlurDateInput.test.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-// eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import DateInput from 'terra-date-input';
 import classNames from 'classnames/bind';
 import styles from './common/DateInput.test.module.scss';