+++ packages/terra-application-links/CHANGELOG.md
@@ -5,6 +5,8 @@ Unreleased
 ----------
 ### Changed
 * updated `master` references to `main`.
+* Minor file changes to comply with eslint rules
+* Remove unused disable eslint directives
 
 6.36.0 - (June 16, 2020)
 ------------------