+++ packages/terra-application-links/CHANGELOG.md
@@ -4,6 +4,13 @@ Changelog
 Unreleased
 ----------
 
+6.37.0 - (July 14, 2020)
+------------------
+### Changed
+* updated `master` references to `main`.
+* Minor file changes to comply with eslint rules
+* Remove unused disable eslint directives
+
 6.36.0 - (June 16, 2020)
 ------------------
 ### Changed